requirements.txt
pathlib
requests
python-dotenv
pytest
random